The COMPO family

The function COMPO() defines the COM-POISSON distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss().

COMPO(mu.link = "log", sigma.link = "log")

Arguments

mu.link

defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link

defines the sigma.link, with "log" link as the default for the sigma.

Value

Returns a gamlss.family object which can be used to fit a COMPO distribution in the gamlss() function.

Details

The COMPO distribution with parameters \(\mu\) and \(\sigma\) has a support 0, 1, 2, ... and mass function given by

\(f(x | \mu, \sigma) = \frac{\mu^x}{(x!)^{\sigma} Z(\mu, \sigma)} \)

with \(\mu > 0\), \(\sigma > 0\) and

\(Z(\mu, \sigma)=\sum_{j=0}^{\infty} \frac{\mu^j}{(j!)^\sigma}\).

The proposed functions here are based on the functions from the COMPoissonReg package.

References

Shmueli, G., Minka, T. P., Kadane, J. B., Borle, S., & Boatwright, P. (2005). A useful distribution for fitting discrete data: revival of the Conway–Maxwell–Poisson distribution. Journal of the Royal Statistical Society Series C: Applied Statistics, 54(1), 127-142.

See also

Author

Freddy Hernandez, fhernanb@unal.edu.co

Examples

# Example 1
# Generating some random values with
# known mu and sigma
set.seed(1234)
y <- rCOMPO(n=100, mu=5, sigma=1.5)

estim_mu_sigma_COMPO(y)
#> [1] 4.402069 1.509795

# Fitting the model
library(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=COMPO,
               control=gamlss.control(n.cyc=500, trace=TRUE))
#> GAMLSS-RS iteration 1: Global Deviance = 335.268 
#> GAMLSS-RS iteration 2: Global Deviance = 335.2722 
#> GAMLSS-RS iteration 3: Global Deviance = 335.2765 
#> GAMLSS-RS iteration 4: Global Deviance = 335.2808 
#> GAMLSS-RS iteration 5: Global Deviance = 335.2849 
#> GAMLSS-RS iteration 6: Global Deviance = 335.289 
#> GAMLSS-RS iteration 7: Global Deviance = 335.293 
#> GAMLSS-RS iteration 8: Global Deviance = 335.2969 
#> GAMLSS-RS iteration 9: Global Deviance = 335.3006 
#> GAMLSS-RS iteration 10: Global Deviance = 335.3041 
#> GAMLSS-RS iteration 11: Global Deviance = 335.3076 
#> GAMLSS-RS iteration 12: Global Deviance = 335.3108 
#> GAMLSS-RS iteration 13: Global Deviance = 335.3139 
#> GAMLSS-RS iteration 14: Global Deviance = 335.3168 
#> GAMLSS-RS iteration 15: Global Deviance = 335.3196 
#> GAMLSS-RS iteration 16: Global Deviance = 335.3222 
#> GAMLSS-RS iteration 17: Global Deviance = 335.3247 
#> GAMLSS-RS iteration 18: Global Deviance = 335.327 
#> GAMLSS-RS iteration 19: Global Deviance = 335.3292 
#> GAMLSS-RS iteration 20: Global Deviance = 335.3313 
#> GAMLSS-RS iteration 21: Global Deviance = 335.3332 
#> GAMLSS-RS iteration 22: Global Deviance = 335.335 
#> GAMLSS-RS iteration 23: Global Deviance = 335.3367 
#> GAMLSS-RS iteration 24: Global Deviance = 335.3383 
#> GAMLSS-RS iteration 25: Global Deviance = 335.3398 
#> GAMLSS-RS iteration 26: Global Deviance = 335.3412 
#> GAMLSS-RS iteration 27: Global Deviance = 335.3425 
#> GAMLSS-RS iteration 28: Global Deviance = 335.3437 
#> GAMLSS-RS iteration 29: Global Deviance = 335.3448 
#> GAMLSS-RS iteration 30: Global Deviance = 335.3459 
#> GAMLSS-RS iteration 31: Global Deviance = 335.347 
#> GAMLSS-RS iteration 32: Global Deviance = 335.3479 

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what="mu"))
#> (Intercept) 
#>    4.228954 
exp(coef(mod1, what="sigma"))
#> (Intercept) 
#>    1.467928 

# Example 2
# Generating random values under some model

# \donttest{
# A function to simulate a data set with Y ~ COMPO
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu    <- exp(2 + 1 * x1) # 12 approximately
  sigma <- exp(2 - 2 * x2) # 2.71 approximately
  y <- rCOMPO(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(123)
dat <- gendat(n=100)

# Fitting the model
mod2 <- NULL
mod2 <- gamlss(y~x1, sigma.fo=~x2, family=COMPO, data=dat,
               control=gamlss.control(n.cyc=500, trace=TRUE))
#> GAMLSS-RS iteration 1: Global Deviance = 359.5696 
#> GAMLSS-RS iteration 2: Global Deviance = 355.8301 
#> GAMLSS-RS iteration 3: Global Deviance = 352.3682 
#> GAMLSS-RS iteration 4: Global Deviance = 349.0032 
#> GAMLSS-RS iteration 5: Global Deviance = 345.587 
#> GAMLSS-RS iteration 6: Global Deviance = 342.2927 
#> GAMLSS-RS iteration 7: Global Deviance = 339.0862 
#> GAMLSS-RS iteration 8: Global Deviance = 335.9689 
#> GAMLSS-RS iteration 9: Global Deviance = 332.9319 
#> GAMLSS-RS iteration 10: Global Deviance = 330.03 
#> GAMLSS-RS iteration 11: Global Deviance = 327.0244 
#> GAMLSS-RS iteration 12: Global Deviance = 324.3343 
#> GAMLSS-RS iteration 13: Global Deviance = 321.7411 
#> GAMLSS-RS iteration 14: Global Deviance = 319.178 
#> GAMLSS-RS iteration 15: Global Deviance = 316.8021 
#> GAMLSS-RS iteration 16: Global Deviance = 314.5448 
#> GAMLSS-RS iteration 17: Global Deviance = 312.3774 
#> GAMLSS-RS iteration 18: Global Deviance = 310.3075 
#> GAMLSS-RS iteration 19: Global Deviance = 308.2979 
#> GAMLSS-RS iteration 20: Global Deviance = 306.4132 
#> GAMLSS-RS iteration 21: Global Deviance = 304.6026 
#> GAMLSS-RS iteration 22: Global Deviance = 302.86 
#> GAMLSS-RS iteration 23: Global Deviance = 300.9943 
#> GAMLSS-RS iteration 24: Global Deviance = 299.2266 
#> GAMLSS-RS iteration 25: Global Deviance = 297.7815 
#> GAMLSS-RS iteration 26: Global Deviance = 296.4056 
#> GAMLSS-RS iteration 27: Global Deviance = 295.1029 
#> GAMLSS-RS iteration 28: Global Deviance = 293.8743 
#> GAMLSS-RS iteration 29: Global Deviance = 292.736 
#> GAMLSS-RS iteration 30: Global Deviance = 291.6615 
#> GAMLSS-RS iteration 31: Global Deviance = 290.6119 
#> GAMLSS-RS iteration 32: Global Deviance = 289.5895 
#> GAMLSS-RS iteration 33: Global Deviance = 288.6284 
#> GAMLSS-RS iteration 34: Global Deviance = 287.7196 
#> GAMLSS-RS iteration 35: Global Deviance = 286.8588 
#> GAMLSS-RS iteration 36: Global Deviance = 286.0207 
#> GAMLSS-RS iteration 37: Global Deviance = 285.2355 
#> GAMLSS-RS iteration 38: Global Deviance = 284.4838 
#> GAMLSS-RS iteration 39: Global Deviance = 283.801 
#> GAMLSS-RS iteration 40: Global Deviance = 283.1456 
#> GAMLSS-RS iteration 41: Global Deviance = 282.5427 
#> GAMLSS-RS iteration 42: Global Deviance = 281.9835 
#> GAMLSS-RS iteration 43: Global Deviance = 281.464 
#> GAMLSS-RS iteration 44: Global Deviance = 280.9805 
#> GAMLSS-RS iteration 45: Global Deviance = 280.5295 
#> GAMLSS-RS iteration 46: Global Deviance = 280.1082 
#> GAMLSS-RS iteration 47: Global Deviance = 279.7138 
#> GAMLSS-RS iteration 48: Global Deviance = 279.3441 
#> GAMLSS-RS iteration 49: Global Deviance = 278.9969 
#> GAMLSS-RS iteration 50: Global Deviance = 278.6702 
#> GAMLSS-RS iteration 51: Global Deviance = 278.3619 
#> GAMLSS-RS iteration 52: Global Deviance = 278.066 
#> GAMLSS-RS iteration 53: Global Deviance = 277.782 
#> GAMLSS-RS iteration 54: Global Deviance = 277.5154 
#> GAMLSS-RS iteration 55: Global Deviance = 277.2632 
#> GAMLSS-RS iteration 56: Global Deviance = 277.0242 
#> GAMLSS-RS iteration 57: Global Deviance = 276.7813 
#> GAMLSS-RS iteration 58: Global Deviance = 276.487 
#> GAMLSS-RS iteration 59: Global Deviance = 276.2235 
#> GAMLSS-RS iteration 60: Global Deviance = 275.9784 
#> GAMLSS-RS iteration 61: Global Deviance = 275.8963 
#> GAMLSS-RS iteration 62: Global Deviance = 275.6844 
#> GAMLSS-RS iteration 63: Global Deviance = 275.4609 
#> GAMLSS-RS iteration 64: Global Deviance = 275.2366 
#> GAMLSS-RS iteration 65: Global Deviance = 275.1128 
#> GAMLSS-RS iteration 66: Global Deviance = 275.0367 
#> GAMLSS-RS iteration 67: Global Deviance = 274.9362 
#> GAMLSS-RS iteration 68: Global Deviance = 274.8324 
#> GAMLSS-RS iteration 69: Global Deviance = 274.7272 
#> GAMLSS-RS iteration 70: Global Deviance = 274.6215 
#> GAMLSS-RS iteration 71: Global Deviance = 274.516 
#> GAMLSS-RS iteration 72: Global Deviance = 274.4111 
#> GAMLSS-RS iteration 73: Global Deviance = 274.3075 
#> GAMLSS-RS iteration 74: Global Deviance = 274.2056 
#> GAMLSS-RS iteration 75: Global Deviance = 274.1059 
#> GAMLSS-RS iteration 76: Global Deviance = 274.0086 
#> GAMLSS-RS iteration 77: Global Deviance = 273.914 
#> GAMLSS-RS iteration 78: Global Deviance = 273.8225 
#> GAMLSS-RS iteration 79: Global Deviance = 273.734 
#> GAMLSS-RS iteration 80: Global Deviance = 273.6487 
#> GAMLSS-RS iteration 81: Global Deviance = 273.5769 
#> GAMLSS-RS iteration 82: Global Deviance = 273.51 
#> GAMLSS-RS iteration 83: Global Deviance = 273.445 
#> GAMLSS-RS iteration 84: Global Deviance = 273.3821 
#> GAMLSS-RS iteration 85: Global Deviance = 273.3212 
#> GAMLSS-RS iteration 86: Global Deviance = 273.2624 
#> GAMLSS-RS iteration 87: Global Deviance = 273.2056 
#> GAMLSS-RS iteration 88: Global Deviance = 273.1508 
#> GAMLSS-RS iteration 89: Global Deviance = 273.098 
#> GAMLSS-RS iteration 90: Global Deviance = 273.0472 
#> GAMLSS-RS iteration 91: Global Deviance = 272.9982 
#> GAMLSS-RS iteration 92: Global Deviance = 272.9511 
#> GAMLSS-RS iteration 93: Global Deviance = 272.9059 
#> GAMLSS-RS iteration 94: Global Deviance = 272.8624 
#> GAMLSS-RS iteration 95: Global Deviance = 272.8206 
#> GAMLSS-RS iteration 96: Global Deviance = 272.7805 
#> GAMLSS-RS iteration 97: Global Deviance = 272.7419 
#> GAMLSS-RS iteration 98: Global Deviance = 272.705 
#> GAMLSS-RS iteration 99: Global Deviance = 272.6695 
#> GAMLSS-RS iteration 100: Global Deviance = 272.6354 
#> GAMLSS-RS iteration 101: Global Deviance = 272.6106 
#> GAMLSS-RS iteration 102: Global Deviance = 272.5873 
#> GAMLSS-RS iteration 103: Global Deviance = 272.5647 
#> GAMLSS-RS iteration 104: Global Deviance = 272.5428 
#> GAMLSS-RS iteration 105: Global Deviance = 272.5214 
#> GAMLSS-RS iteration 106: Global Deviance = 272.5007 
#> GAMLSS-RS iteration 107: Global Deviance = 272.4805 
#> GAMLSS-RS iteration 108: Global Deviance = 272.4609 
#> GAMLSS-RS iteration 109: Global Deviance = 272.4419 
#> GAMLSS-RS iteration 110: Global Deviance = 272.4234 
#> GAMLSS-RS iteration 111: Global Deviance = 272.4053 
#> GAMLSS-RS iteration 112: Global Deviance = 272.3878 
#> GAMLSS-RS iteration 113: Global Deviance = 272.3708 
#> GAMLSS-RS iteration 114: Global Deviance = 272.3543 
#> GAMLSS-RS iteration 115: Global Deviance = 272.3382 
#> GAMLSS-RS iteration 116: Global Deviance = 272.3226 
#> GAMLSS-RS iteration 117: Global Deviance = 272.3074 
#> GAMLSS-RS iteration 118: Global Deviance = 272.2927 
#> GAMLSS-RS iteration 119: Global Deviance = 272.2784 
#> GAMLSS-RS iteration 120: Global Deviance = 272.2644 
#> GAMLSS-RS iteration 121: Global Deviance = 272.2509 
#> GAMLSS-RS iteration 122: Global Deviance = 272.2378 
#> GAMLSS-RS iteration 123: Global Deviance = 272.2251 
#> GAMLSS-RS iteration 124: Global Deviance = 272.2127 
#> GAMLSS-RS iteration 125: Global Deviance = 272.2007 
#> GAMLSS-RS iteration 126: Global Deviance = 272.189 
#> GAMLSS-RS iteration 127: Global Deviance = 272.1777 
#> GAMLSS-RS iteration 128: Global Deviance = 272.1666 
#> GAMLSS-RS iteration 129: Global Deviance = 272.1559 
#> GAMLSS-RS iteration 130: Global Deviance = 272.1455 
#> GAMLSS-RS iteration 131: Global Deviance = 272.1354 
#> GAMLSS-RS iteration 132: Global Deviance = 272.1256 
#> GAMLSS-RS iteration 133: Global Deviance = 272.1161 
#> GAMLSS-RS iteration 134: Global Deviance = 272.1069 
#> GAMLSS-RS iteration 135: Global Deviance = 272.0979 
#> GAMLSS-RS iteration 136: Global Deviance = 272.0892 
#> GAMLSS-RS iteration 137: Global Deviance = 272.0807 
#> GAMLSS-RS iteration 138: Global Deviance = 272.0725 
#> GAMLSS-RS iteration 139: Global Deviance = 272.0645 
#> GAMLSS-RS iteration 140: Global Deviance = 272.0567 
#> GAMLSS-RS iteration 141: Global Deviance = 272.0492 
#> GAMLSS-RS iteration 142: Global Deviance = 272.0418 
#> GAMLSS-RS iteration 143: Global Deviance = 272.0347 
#> GAMLSS-RS iteration 144: Global Deviance = 272.0278 
#> GAMLSS-RS iteration 145: Global Deviance = 272.0211 
#> GAMLSS-RS iteration 146: Global Deviance = 272.0072 
#> GAMLSS-RS iteration 147: Global Deviance = 271.9942 
#> GAMLSS-RS iteration 148: Global Deviance = 271.9821 
#> GAMLSS-RS iteration 149: Global Deviance = 271.9708 
#> GAMLSS-RS iteration 150: Global Deviance = 271.9603 
#> GAMLSS-RS iteration 151: Global Deviance = 271.9502 
#> GAMLSS-RS iteration 152: Global Deviance = 271.9407 
#> GAMLSS-RS iteration 153: Global Deviance = 271.9318 
#> GAMLSS-RS iteration 154: Global Deviance = 271.9235 
#> GAMLSS-RS iteration 155: Global Deviance = 271.9157 
#> GAMLSS-RS iteration 156: Global Deviance = 271.9083 
#> GAMLSS-RS iteration 157: Global Deviance = 271.9014 
#> GAMLSS-RS iteration 158: Global Deviance = 271.8949 
#> GAMLSS-RS iteration 159: Global Deviance = 271.8888 
#> GAMLSS-RS iteration 160: Global Deviance = 271.8831 
#> GAMLSS-RS iteration 161: Global Deviance = 271.8776 
#> GAMLSS-RS iteration 162: Global Deviance = 271.8725 
#> GAMLSS-RS iteration 163: Global Deviance = 271.8677 
#> GAMLSS-RS iteration 164: Global Deviance = 271.8631 
#> GAMLSS-RS iteration 165: Global Deviance = 271.8588 
#> GAMLSS-RS iteration 166: Global Deviance = 271.8548 
#> GAMLSS-RS iteration 167: Global Deviance = 271.8509 
#> GAMLSS-RS iteration 168: Global Deviance = 271.8473 
#> GAMLSS-RS iteration 169: Global Deviance = 271.8439 
#> GAMLSS-RS iteration 170: Global Deviance = 271.8407 
#> GAMLSS-RS iteration 171: Global Deviance = 271.8376 
#> GAMLSS-RS iteration 172: Global Deviance = 271.8347 
#> GAMLSS-RS iteration 173: Global Deviance = 271.832 
#> GAMLSS-RS iteration 174: Global Deviance = 271.8294 
#> GAMLSS-RS iteration 175: Global Deviance = 271.827 
#> GAMLSS-RS iteration 176: Global Deviance = 271.8246 
#> GAMLSS-RS iteration 177: Global Deviance = 271.8225 
#> GAMLSS-RS iteration 178: Global Deviance = 271.8204 
#> GAMLSS-RS iteration 179: Global Deviance = 271.8184 
#> GAMLSS-RS iteration 180: Global Deviance = 271.8166 
#> GAMLSS-RS iteration 181: Global Deviance = 271.8148 
#> GAMLSS-RS iteration 182: Global Deviance = 271.8131 
#> GAMLSS-RS iteration 183: Global Deviance = 271.8116 
#> GAMLSS-RS iteration 184: Global Deviance = 271.8101 
#> GAMLSS-RS iteration 185: Global Deviance = 271.8086 
#> GAMLSS-RS iteration 186: Global Deviance = 271.8073 
#> GAMLSS-RS iteration 187: Global Deviance = 271.806 
#> GAMLSS-RS iteration 188: Global Deviance = 271.8048 
#> GAMLSS-RS iteration 189: Global Deviance = 271.8036 
#> GAMLSS-RS iteration 190: Global Deviance = 271.8026 
#> GAMLSS-RS iteration 191: Global Deviance = 271.8015 
#> GAMLSS-RS iteration 192: Global Deviance = 271.8005 

summary(mod2)
#> Warning: summary: vcov has failed, option qr is used instead
#> ******************************************************************
#> Family:  c("COMPO", "Conway-Maxwell-Poisson") 
#> 
#> Call:  gamlss(formula = y ~ x1, sigma.formula = ~x2, family = COMPO,  
#>     data = dat, control = gamlss.control(n.cyc = 500, trace = TRUE)) 
#> 
#> Fitting method: RS() 
#> 
#> ------------------------------------------------------------------
#> Mu link function:  log
#> Mu Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   2.1595     0.1702  12.689  < 2e-16 ***
#> x1            1.0749     0.2950   3.644 0.000432 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ------------------------------------------------------------------
#> Sigma link function:  log
#> Sigma Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  2.01726    0.08954   22.53   <2e-16 ***
#> x2          -1.87873    0.12080  -15.55   <2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ------------------------------------------------------------------
#> No. of observations in the fit:  100 
#> Degrees of Freedom for the fit:  4
#>       Residual Deg. of Freedom:  96 
#>                       at cycle:  192 
#>  
#> Global Deviance:     271.8005 
#>             AIC:     279.8005 
#>             SBC:     290.2212 
#> ******************************************************************
# }