The Discrete Generalized Burr Hatke family

The function DBH() defines the Discrete Burr Hatke distribution, one-parameter discrete distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss().

DGBH(mu.link = "logit")

Arguments

mu.link

defines the mu.link, with "logit" link as the default for the mu parameter. Other links are "probit" and "cloglog"'(complementary log-log)

Details

The Discrete Burr-Hatke distribution with parameters \(\mu\) has a support 0, 1, 2, ... and density given by

\(f(x | \mu) = (\frac{1}{x+1}-\frac{\mu}{x+2})\mu^{x}\)

The pmf is log-convex for all values of \(0 < \mu < 1\), where \(\frac{f(x+1;\mu)}{f(x;\mu)}\) is an increasing function in \(x\) for all values of the parameter \(\mu\).

Note: in this implementation we changed the original parameters \(\lambda\) for \(\mu\), we did it to implement this distribution within gamlss framework.

References

Yousof HM, Chesneau C, Hamedani G, Ibrahim M (2021). “A new discrete distribution: Properties, characterizations, modeling real count data, Bayesian and non-Bayesian estimations.” Statistica.

See also

Author

Sofia Cuartas García, scuartasg@unal.edu.co

Examples

# Example 1
# Generating some random values with
# known mu and sigma
y <- rDGBH(n=10, mu=0.65, sigma=1.5)
y
#>  [1] 0 0 1 0 0 0 1 0 0 0

# Example 2
# Generating random values under some model