The discrete Inverted Kumaraswamy family

The function DIKUM() defines the discrete Inverted Kumaraswamy distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss().

DIKUM(mu.link = "log", sigma.link = "log")

Arguments

mu.link

defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link

defines the sigma.link, with "log" link as the default for the sigma.

Details

The discrete Inverted Kumaraswamy distribution with parameters \(\mu\) and \(\sigma\) has a support 0, 1, 2, ... and density given by

\(f(x | \mu, \sigma) = (1-(2+x)^{-\mu})^{\sigma}-(1-(1+x)^{-\mu})^{\sigma}\)

with \(\mu > 0\) and \(\sigma > 0\).

Note: in this implementation we changed the original parameters \(\alpha\) and \(\beta\) for \(\mu\) and \(\sigma\) respectively, we did it to implement this distribution within gamlss framework.

References

EL-Helbawy AA, Hegazy MA, AL-Dayian GR, Abd EL-Kader RE (2022). “A Discrete Analog of the Inverted Kumaraswamy Distribution: Properties and Estimation with Application to COVID-19 Data.” Pakistan Journal of Statistics & Operation Research, 18(1).

See also

Author

Daniel Felipe Villa Rengifo, dvilla@unal.edu.co

Examples

# Example 1
# Generating some random values with
# known mu and sigma
set.seed(150)
y <- rDIKUM(1000, mu=1, sigma=5)

# Fitting the model
library(gamlss)

mod1 <- gamlss(y ~ 1, sigma.fo = ~1, family=DIKUM,
               control = gamlss.control(n.cyc=500, trace=FALSE))

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what='mu'))
#> (Intercept) 
#>   0.9977962 
exp(coef(mod1, what='sigma'))
#> (Intercept) 
#>    4.955013 

# Example 2
# Generating random values under some model

library(gamlss)

# A function to simulate a data set with Y ~ DIKUM
gendat <- function(n) {
  x1 <- runif(n, min=0.4, max=0.6)
  x2 <- runif(n, min=0.4, max=0.6)
  mu    <- exp(1.21 - 3 * x1) # 0.75 en promedio
  sigma <- exp(1.26 - 2 * x2) # 1.30 en promedio
  y <- rDIKUM(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

dat <- gendat(n=150)

# Fitting the model
mod2 <- gamlss(y ~ x1, sigma.fo = ~x2, family = "DIKUM", data=dat,
               control=gamlss.control(n.cyc=500, trace=FALSE))

summary(mod2)
#> Warning: summary: vcov has failed, option qr is used instead
#> ******************************************************************
#> Family:  c("DIKUM", "discrete-Inverted-Kumaraswamy") 
#> 
#> Call:  gamlss(formula = y ~ x1, sigma.formula = ~x2, family = "DIKUM",  
#>     data = dat, control = gamlss.control(n.cyc = 500, trace = FALSE)) 
#> 
#> Fitting method: RS() 
#> 
#> ------------------------------------------------------------------
#> Mu link function:  log
#> Mu Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)  
#> (Intercept)   1.0937     0.4966   2.202   0.0292 *
#> x1           -2.2099     0.9829  -2.248   0.0260 *
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ------------------------------------------------------------------
#> Sigma link function:  log
#> Sigma Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   3.4109     0.8629   3.953 0.000119 ***
#> x2           -5.4919     1.7364  -3.163 0.001896 ** 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> ------------------------------------------------------------------
#> No. of observations in the fit:  150 
#> Degrees of Freedom for the fit:  4
#>       Residual Deg. of Freedom:  146 
#>                       at cycle:  11 
#>  
#> Global Deviance:     851.1499 
#>             AIC:     859.1499 
#>             SBC:     871.1924 
#> ******************************************************************