The hyper Poisson family (with mu as mean)

The function HYPERPO2() defines the hyper Poisson distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss().

HYPERPO2(mu.link = "log", sigma.link = "log")

Arguments

mu.link

defines the mu.link, with "log" link as the default for the mu parameter.

sigma.link

defines the sigma.link, with "log" link as the default for the sigma.

Details

The hyper-Poisson distribution with parameters \(\mu\) and \(\sigma\) has a support 0, 1, 2, ...

Note: in this implementation the parameter \(\mu\) is the mean of the distribution and \(\sigma\) corresponds to the dispersion parameter. If you fit a model with this parameterization, the time will increase because an internal procedure to convert \(\mu\) to \(\lambda\) parameter.

References

Sáez-Castillo AJ, Conde-Sánchez A (2013). “A hyper-Poisson regression model for overdispersed and underdispersed count data.” Computational Statistics & Data Analysis, 61, 148--157.

See also

Examples

# Example 1
# Generating some random values with
# known mu and sigma
y <- rHYPERPO2(n=100, mu=5, sigma=1.5)

# Fitting the model
library(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=HYPERPO2,
               control=gamlss.control(n.cyc=500, trace=FALSE))

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what='mu'))
#> (Intercept) 
#>    5.185234 
exp(coef(mod1, what='sigma'))
#> (Intercept) 
#>   0.8867399 

# Example 2
# Generating random values under some model

if (FALSE) {
  # A function to simulate a data set with Y ~ HYPERPO
  gendat <- function(n) {
    x1 <- runif(n)
    x2 <- runif(n)
    mu    <- exp(1.21 - 3 * x1) # 0.75 en promedio
    sigma <- exp(1.26 - 2 * x2) # 1.30 en promedio
    y <- rHYPERPO2(n=n, mu=mu, sigma=sigma)
    data.frame(y=y, x1=x1,x2=x2)
  }

  set.seed(1235)
  datos <- gendat(n=150)

  mod2 <- NULL
  mod2 <- gamlss(y~x1, sigma.fo=~x2, family=HYPERPO2, data=datos,
                 control=gamlss.control(n.cyc=500, trace=TRUE))

  summary(mod2)
}