The Discrete Generalized Burr Hatke distribution

These functions define the density, distribution function, quantile function and random generation for the Discrete Generalized Burr Hatke distribution with parameter \(\mu\).

dDGBH(x, mu = 0.999, sigma = 100, log = FALSE)

pDGBH(q, mu = 0.999, sigma = 100, lower.tail = TRUE, log.p = FALSE)

qDGBH(p, mu = mu, sigma = sigma, lower.tail = TRUE)

rDGBH(n, mu = 0.999, sigma = 100)

Arguments

x, q

vector of (non-negative integer) quantiles.

mu

vector of the mu parameter.

sigma

vector of the sigma parameter.

log, log.p

logical; if TRUE, probabilities p are given as log(p).

lower.tail

logical; if TRUE (default), probabilities are \(P[X <= x]\), otherwise, P[X > x].

p

vector of probabilities.

n

number of random values to return.

Value

dDGBH gives the density, pDGBH gives the distribution function, qDGBH gives the quantile function, rDGBH

generates random deviates.

Details

The Discrete Burr-Hatke distribution with parameters \(\mu\) has a support 0, 1, 2, ... and density given by

\(f(x | \mu) = (\frac{1}{x+1}-\frac{\mu}{x+2})\mu^{x}\)

The pmf is log-convex for all values of \(0 < \mu < 1\), where \(\frac{f(x+1;\mu)}{f(x;\mu)}\) is an increasing function in \(x\) for all values of the parameter \(\mu\).

Note: in this implementation we changed the original parameters \(\lambda\) for \(\mu\), we did it to implement this distribution within gamlss framework.

References

Yousof HM, Chesneau C, Hamedani G, Ibrahim M (2021). “A new discrete distribution: Properties, characterizations, modeling real count data, Bayesian and non-Bayesian estimations.” Statistica.

See also

Author

Sofia Cuartas García, scuartasg@unal.edu.co

Examples

# Example 1
dDGBH(x=0:10, mu=0.56, sigma=1.1)
#>  [1] 7.200000e-01 1.838149e-01 6.031108e-02 2.194229e-02 8.399111e-03
#>  [6] 3.306523e-03 1.323357e-03 5.349608e-04 2.175673e-04 8.879722e-05
#> [11] 3.630897e-05

# Example 2
# Checking if curves go to 1

# Example 3
# Some tests to compare the simulator with theoretical probabilities

# Example 4