The Discrete Burr Hatke family

The function DBH() defines the Discrete Burr Hatke distribution, one-parameter discrete distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss().

DBH(mu.link = "logit")

Arguments

mu.link

defines the mu.link, with "logit" link as the default for the mu parameter. Other links are "probit" and "cloglog"'(complementary log-log)

Details

The Discrete Burr-Hatke distribution with parameters \(\mu\) has a support 0, 1, 2, ... and density given by

\(f(x | \mu) = (\frac{1}{x+1}-\frac{\mu}{x+2})\mu^{x}\)

The pmf is log-convex for all values of \(0 < \mu < 1\), where \(\frac{f(x+1;\mu)}{f(x;\mu)}\) is an increasing function in \(x\) for all values of the parameter \(\mu\).

Note: in this implementation we changed the original parameters \(\lambda\) for \(\mu\), we did it to implement this distribution within gamlss framework.

References

El-Morshedy M, Eliwa MS, Altun E (2020). “Discrete Burr-Hatke distribution with properties, estimation methods and regression model.” IEEE access, 8, 74359--74370.

See also

Author

Valentina Hurtado Sepulveda, vhurtados@unal.edu.co

Examples

# Example 1
# Generating some random values with
# known mu and sigma
y <- rHYPERPO(n=300, mu=5, sigma=1.5)

# Fitting the model
library(gamlss)
#> Loading required package: splines
#> Loading required package: gamlss.data
#> 
#> Attaching package: 'gamlss.data'
#> The following object is masked from 'package:datasets':
#> 
#>     sleep
#> Loading required package: gamlss.dist
#> Loading required package: MASS
#> Loading required package: nlme
#> Loading required package: parallel
#>  **********   GAMLSS Version 5.4-10  ********** 
#> For more on GAMLSS look at https://www.gamlss.com/
#> Type gamlssNews() to see new features/changes/bug fixes.
mod1 <- gamlss(y~1, sigma.fo=~1, family=HYPERPO,
               control=gamlss.control(n.cyc=500, trace=FALSE))

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what='mu'))
#> (Intercept) 
#>    4.843747 
exp(coef(mod1, what='sigma'))
#> (Intercept) 
#>    1.304309 

# Example 2
# Generating random values under some model

if (FALSE) {
  # A function to simulate a data set with Y ~ HYPERPO
  gendat <- function(n) {
    x1 <- runif(n)
    x2 <- runif(n)
    mu    <- exp(1.21 - 3 * x1) # 0.75 en promedio
    sigma <- exp(1.26 - 2 * x2) # 1.30 en promedio
    y <- rHYPERPO(n=n, mu=mu, sigma=sigma)
    data.frame(y=y, x1=x1,x2=x2)
  }

  set.seed(1235)
  datos <- gendat(n=150)

  mod2 <- NULL
  mod2 <- gamlss(y~x1, sigma.fo=~x2, family=HYPERPO, data=datos,
                 control=gamlss.control(n.cyc=500, trace=TRUE))

  summary(mod2)
}